update models, fixed problems with rendering

+++ Models/Card.cs
@@ -7,11 +7,11 @@ namespace KanbanRby.Models;
 public class Card :  BaseModel
 {
     [PrimaryKey("id")]
-    public long Id { get; set; }
+    public int Id { get; set; }
     
     [Reference(typeof(Column))]
     [Column("column_id")]
-    public long ColumnId { get; set; }
+    public int ColumnId { get; set; }
     
     [Column("name")]
     public string Name { get; set; }
