+++ Models/Kanban.cs
@@ -7,10 +7,10 @@ namespace KanbanRby.Models;
 public class Kanban : BaseModel
 {
     [PrimaryKey("id")]
-    public long Id { get; set; }
+    public int Id { get; set; }
     
     [Column("user_id")]
-    public long? UserId { get; set; }
+    public int? UserId { get; set; }
     
     [Column("name")]
     public string? Name { get; set; }
